# `MessageTemplateParser` refactor incl. experiemental `.` support