--- a/.gitignore
+++ b/.gitignore
@@ -331,3 +331,5 @@ ASALocalRun/
 *.orig
 
 *.received.txt
+
+.DS_Store