--- a/test/Serilog.Tests/Parsing/MessageTemplateParserTests.cs
+++ b/test/Serilog.Tests/Parsing/MessageTemplateParserTests.cs
@@ -14,8 +14,8 @@ public class MessageTemplateParserTests
     {
         var parsed = Parse(message);
         Assert.Equal(
-            parsed,
-            messageTemplateTokens);
+            messageTemplateTokens,
+            parsed);
     }
 
     [Fact]