@nblumhardt 2024-05-17
(Surprising actually, that this late in the game a couple of simple `string.IndexOf()` calls can > 3x performance and < 1/3x allocations in a fairly mainline scenario 😅 .)