@nblumhardt 2024-05-18
Heads-up also @Mpdreamz - I know that `Elastic.Serilog.Sinks` currently maps some `PascalCase` property names to dotted ECS equivalents. It'd be interesting to consider what a future iteration supporting dotted property names in message templates directly would mean over there.