@nblumhardt 2024-05-19
Updated the PR to consider `.` only, and ignore `-` for now.