@Mpdreamz 2024-05-21
Love this! Our ECS field based setters can handle both the original dotted key names as well as PascalCased ones.