@cchapman-kxs 2024-05-21
Also verified simple console and file logging works as expected.  Sending to DataDog is using the `DurableHttpUsingFileSizeRolledBuffers` from https://github.com/FantasticFiasco/serilog-sinks-http - very much not a trivial example and many places that could have interfered with the intended result.