@nblumhardt 2024-11-25
Thinking about opening a PR to stabilize this. Any experiences to share, good or bad?